Fix missed cast to ushort on prior commit.

+++ Rubberduck.Core/Common/RubberduckHooks.cs
@@ -172,7 +172,7 @@ namespace Rubberduck.Common
             var processed = false;
             try
             {
-                var hook = Hooks.OfType<Hotkey>().SingleOrDefault(k => k.HotkeyInfo.HookId == wParam);
+                var hook = Hooks.OfType<Hotkey>().SingleOrDefault(k => k.HotkeyInfo.HookId == (ushort)wParam);
                 if (hook != null)
                 {
                     hook.OnMessageReceived();
